+++ sample-apps/react-native/cookbook/src/recipes/custom-participants-layout/CustomParticipantsLayoutScreen.tsx
@@ -1,5 +1,5 @@
-import {ActivityIndicator, StyleSheet, View} from 'react-native';
-import React, {useEffect} from 'react';
+import {ActivityIndicator, Platform, StyleSheet, View} from 'react-native';
+import React, {useEffect, useRef} from 'react';
 import {
   CallControlsView,
   StreamVideoCall,
@@ -40,6 +40,9 @@ export default () => {
     user: USER,
     tokenOrProvider: USER.custom.token,
     apiKey: STREAM_API_KEY,
+    options: {
+      preferredVideoCodec: Platform.OS === 'android' ? 'VP8' : undefined,
+    },
   });
   const handleOnCallHungUp = () => navigation.navigate('WelcomeScreen');