+++ sample-apps/react/react-video-demo/src/components/GiphyPreview/index.ts
@@ -1,0 +1,1 @@
+export { GiphyPreview as default } from './GiphyPreview';